(C++) Compute Orthogonal and Diagonal Least Cost Matrix from a Distance Matrix

Computes the least cost matrix from a distance matrix. Considers diagonals during computation of least-costs.

Usage

cost_matrix_diagonal_cpp(dist_matrix)

Arguments

dist_matrix

(required, distance matrix). Square distance matrix, output of distance_matrix_cpp().

Value

Least cost matrix.

See also