Dissimilarity Model Frame

Generates a model frame from a dissimilarity data frame generated by distantia() and a data frame static descriptors of the time series.

Usage

distantia_model_frame(
  df = NULL,
  predictors_df = NULL,
  predictors_list = NULL,
  distance = "euclidean"
)

Arguments

df

(required, data frame) Output of distantia() or distantia_aggregate(). Default: NULL

predictors_df

(required, data frame) data frame with numeric predictors to be added to the model frame. Must have a column with the names in df$x and df$y. If sf data frame, the predictor "distance" is added to the model frame. Default: NULL

predictors_list

(optional, list) list defining predictors as combinations fo columns in predictors_df. For example, predictors_list = list(distance = c("x", "y")) uses the columns "x" and "y" from predictors_df to generate the predictor distance in the model frame. Can be used to combine groups of predictors with common themes. Any column from predictors_df not in this list is returned as an individual predictor in the model frame. Default: NULL

distance

(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset distances. Default: "euclidean".

Value

data frame

See also

Examples

#covid prevalence in California counties
tsl <- tsl_initialize(
  x = covid_prevalence,
  name_column = "name",
  time_column = "time"
)

#dissimilarity analysis
df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#combine predictors together
predictors_list <- list(
  economy = c(
    "poverty_percentage",
    "median_income",
    "domestic_product"
    )
)

#generate model frame
model_frame <- distantia_model_frame(
  df = df,
  predictors_df = covid_counties,
  predictors_list = predictors_list
)

head(model_frame)
#>              x            y      psi  distance   economy area_hectares
#> 1      Alameda        Butte 2.962963 158650.80 2.7802798   240683.3223
#> 2      Alameda Contra_Costa 1.162055      0.00 0.3641648      408.0386
#> 3        Butte Contra_Costa 2.733068 132046.29 2.8431733   241091.3609
#> 4      Alameda    El_Dorado 2.483755  89748.63 1.2752482   268655.5879
#> 5        Butte    El_Dorado 2.327273  56927.91 2.1431460    27972.2656
#> 6 Contra_Costa    El_Dorado 2.767442  65680.55 1.1381072   269063.6265
#>   population poverty_percentage median_income domestic_product
#> 1    1430602                7.2         45989       40363.9839
#> 2     488541                0.6          1624       15663.7551
#> 3     942061                7.8         47613       24700.2288
#> 4    1447188                0.6         21282       40701.1966
#> 5      16586                7.8         24707         337.2127
#> 6     958647                0.0         22906       25037.4415
#>   daily_miles_traveled employed_percentage
#> 1            0.2383988            6.611775
#> 2            2.4454332            2.471545
#> 3            2.2070344            4.140230
#> 4            0.2925610            0.869032
#> 5            0.5309598            5.742743
#> 6            2.7379942            1.602513