Stats of Dissimilarity Data Frame

Takes the output of distantia() to return a data frame with one row per time series with the stats of its dissimilarity scores with all other time series.

Usage

distantia_stats(df = NULL)

Arguments

df

(required, data frame) Output of distantia(), distantia_ls(), distantia_dtw(), or distantia_time_shift(). Default: NULL

Value

data frame

See also

Examples

tsl <- tsl_simulate(
  n = 5,
  irregular = FALSE
  )

df <- distantia(
  tsl = tsl,
  lock_step = TRUE
  )

df_stats <- distantia_stats(df = df)

df_stats
#>   name     mean      min       q1   median       q3      max        sd    range
#> 1    A 2.822934 1.107088 2.231241 3.054622 3.646315 4.075402 1.2939296 2.968314
#> 2    B 2.986980 1.107088 2.606637 3.346302 3.726645 4.148229 1.3235986 3.041141
#> 3    C 3.525711 2.605959 2.981355 3.428742 3.973098 4.639400 0.8780187 2.033441
#> 4    D 4.635603 4.075402 4.130022 4.393814 4.899396 5.679382 0.7395604 1.603980
#> 5    E 4.129946 3.503286 3.565409 3.668557 4.233093 5.679382 1.0380765 2.176096