Pollen Counts of Nine Interglacial Sites in Central Europe

Pollen counts of nine interglacial sites in central Europe.

Site coordinates for this dataset are in eemian_coordinates.

Usage

data(eemian_pollen)

Format

data frame with 24 columns and 376 rows.

See also

Examples

data("eemian_pollen")

#to time series list
tsl <- tsl_initialize(
  x = eemian_pollen,
  name_column = "name",
  time_column = "time"
)
#> distantia::utils_prepare_time():  duplicated time indices in 'Krumbach_I':
#> - value 6.8 replaced with 6.825.

#time series plot
if(interactive()){

 tsl_plot(
  tsl = tsl_subset(
    tsl = tsl,
    names = 1:3
    ),
  columns = 2,
  guide_columns = 2
  )

}