Stats of Dissimilarity Data Frame

Takes the output of distantia() to return a data frame with one row per time series with the stats of its dissimilarity scores with all other time series.

Usage

momentum_stats(df = NULL)

Arguments

df

(required, data frame) Output of distantia() or momentum(). Default: NULL

Value

data frame

See also

Examples

tsl <- tsl_simulate(
  n = 5,
  irregular = FALSE
  )

df <- distantia(
  tsl = tsl,
  lock_step = TRUE
  )

df_stats <- distantia_stats(df = df)

df_stats
#>   name     mean      min       q1   median       q3      max        sd
#> 1    A 3.157952 1.385596 2.462905 3.319535 4.014583 4.607144 1.3890870
#> 2    B 3.246655 1.385596 2.434121 3.414387 4.226920 4.772250 1.4880588
#> 3    C 3.256157 2.783629 2.812413 3.021915 3.465659 4.197169 0.6578884
#> 4    D 4.573732 4.197169 4.504650 4.662754 4.731836 4.772250 0.2602843
#> 5    E 3.950598 3.221823 3.668252 3.931103 4.213449 4.718365 0.6184254
#>       range
#> 1 3.2215480
#> 2 3.3866537
#> 3 1.4135400
#> 4 0.5750804
#> 5 1.4965421