Momentum Data Frame to Wide Format

Transforms a data frame returned by momentum() to wide format with the following columns:

most_similar: name of the variable with the highest contribution to similarity (most negative value in the importance column) for each pair of time series.

most_dissimilar: name of the variable with the highest contribution to dissimilarity (most positive value in the importance column) for each pair of time series.

importance__variable_name: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.

psi_only_with__variable_name: dissimilarity of the two time series when only using the given variable.

psi_without__variable_name: dissimilarity of the two time series when removing the given variable.

Usage

momentum_to_wide(df = NULL, sep = "__")

Arguments

df

(required, data frame) Output of momentum(). Default: NULL

sep

(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".

Value

data frame

See also

Examples

tsl <- tsl_initialize(
  x = distantia::albatross,
  name_column = "name",
  time_column = "time"
) |>
  tsl_transform(
    f = f_scale_global
  )

#importance data frame
df <- momentum(
  tsl = tsl
)

df
#>       x    y      psi    variable   importance               effect
#> 1  X132 X134 1.299380           x   87.6066043 decreases similarity
#> 2  X132 X134 1.299380           y   93.9587187 decreases similarity
#> 3  X132 X134 1.299380       speed  -21.9171171 increases similarity
#> 4  X132 X134 1.299380 temperature   72.8121621 decreases similarity
#> 5  X132 X134 1.299380     heading  -38.0165137 increases similarity
#> 6  X132 X136 2.449381           x   48.3845903 decreases similarity
#> 7  X132 X136 2.449381           y   93.5214543 decreases similarity
#> 8  X132 X136 2.449381       speed  -61.1729252 increases similarity
#> 9  X132 X136 2.449381 temperature  356.8824838 decreases similarity
#> 10 X132 X136 2.449381     heading -102.9830173 increases similarity
#> 11 X132 X153 2.091923           x  427.7381576 decreases similarity
#> 12 X132 X153 2.091923           y  156.1285451 decreases similarity
#> 13 X132 X153 2.091923       speed  -40.9249630 increases similarity
#> 14 X132 X153 2.091923 temperature  -14.2831545 increases similarity
#> 15 X132 X153 2.091923     heading  -79.3532025 increases similarity
#> 16 X134 X136 2.358040           x   61.3361468 decreases similarity
#> 17 X134 X136 2.358040           y  108.9650664 decreases similarity
#> 18 X134 X136 2.358040       speed  -59.2603918 increases similarity
#> 19 X134 X136 2.358040 temperature  310.6812842 decreases similarity
#> 20 X134 X136 2.358040     heading  -90.2797292 increases similarity
#> 21 X134 X153 2.074241           x  592.0783167 decreases similarity
#> 22 X134 X153 2.074241           y  116.4310429 decreases similarity
#> 23 X134 X153 2.074241       speed  -52.4149093 increases similarity
#> 24 X134 X153 2.074241 temperature    0.9936944 decreases similarity
#> 25 X134 X153 2.074241     heading  -85.0271172 increases similarity
#> 26 X136 X153 2.666099           x  507.6153648 decreases similarity
#> 27 X136 X153 2.666099           y   56.6957442 decreases similarity
#> 28 X136 X153 2.666099       speed  -65.4516103 increases similarity
#> 29 X136 X153 2.666099 temperature  240.9053814 decreases similarity
#> 30 X136 X153 2.666099     heading -116.2461929 increases similarity
#>    psi_difference psi_without psi_only_with  distance diagonal bandwidth
#> 1      1.13834302    1.253464      2.391807 euclidean     TRUE         1
#> 2      1.22088115    1.232522      2.453403 euclidean     TRUE         1
#> 3     -0.28478672    1.403873      1.119086 euclidean     TRUE         1
#> 4      0.94610694    1.244877      2.190984 euclidean     TRUE         1
#> 5     -0.49397912    1.569450      1.075471 euclidean     TRUE         1
#> 6      1.18512314    2.425775      3.610898 euclidean     TRUE         1
#> 7      2.29069707    2.353934      4.644631 euclidean     TRUE         1
#> 8     -1.49835823    2.895746      1.397388 euclidean     TRUE         1
#> 9      8.74141303    1.586215     10.327628 euclidean     TRUE         1
#> 10    -2.52244683    3.726621      1.204174 euclidean     TRUE         1
#> 11     8.94795300    1.709049     10.657002 euclidean     TRUE         1
#> 12     3.26608898    1.837195      5.103284 euclidean     TRUE         1
#> 13    -0.85611872    2.334915      1.478797 euclidean     TRUE         1
#> 14    -0.29879260    2.127711      1.828918 euclidean     TRUE         1
#> 15    -1.66000791    3.033689      1.373681 euclidean     TRUE         1
#> 16     1.44633065    2.332406      3.778736 euclidean     TRUE         1
#> 17     2.56943945    2.270057      4.839496 euclidean     TRUE         1
#> 18    -1.39738352    2.823389      1.426005 euclidean     TRUE         1
#> 19     7.32598781    1.628667      8.954655 euclidean     TRUE         1
#> 20    -2.12883179    3.443185      1.314354 euclidean     TRUE         1
#> 21    12.28112833    1.516331     13.797459 euclidean     TRUE         1
#> 22     2.41505987    1.903884      4.318944 euclidean     TRUE         1
#> 23    -1.08721129    2.408284      1.321072 euclidean     TRUE         1
#> 24     0.02061161    2.095194      2.115805 euclidean     TRUE         1
#> 25    -1.76366692    3.027265      1.263598 euclidean     TRUE         1
#> 26    13.53352947    2.362495     15.896024 euclidean     TRUE         1
#> 27     1.51156482    2.615352      4.126917 euclidean     TRUE         1
#> 28    -1.74500490    3.130579      1.385574 euclidean     TRUE         1
#> 29     6.42277659    2.007003      8.429779 euclidean     TRUE         1
#> 30    -3.09923889    4.343310      1.244071 euclidean     TRUE         1
#>    lock_step robust
#> 1      FALSE   TRUE
#> 2      FALSE   TRUE
#> 3      FALSE   TRUE
#> 4      FALSE   TRUE
#> 5      FALSE   TRUE
#> 6      FALSE   TRUE
#> 7      FALSE   TRUE
#> 8      FALSE   TRUE
#> 9      FALSE   TRUE
#> 10     FALSE   TRUE
#> 11     FALSE   TRUE
#> 12     FALSE   TRUE
#> 13     FALSE   TRUE
#> 14     FALSE   TRUE
#> 15     FALSE   TRUE
#> 16     FALSE   TRUE
#> 17     FALSE   TRUE
#> 18     FALSE   TRUE
#> 19     FALSE   TRUE
#> 20     FALSE   TRUE
#> 21     FALSE   TRUE
#> 22     FALSE   TRUE
#> 23     FALSE   TRUE
#> 24     FALSE   TRUE
#> 25     FALSE   TRUE
#> 26     FALSE   TRUE
#> 27     FALSE   TRUE
#> 28     FALSE   TRUE
#> 29     FALSE   TRUE
#> 30     FALSE   TRUE

#to wide format
df_wide <- momentum_to_wide(
  df = df
)

df_wide
#>      x    y      psi most_similar most_dissimilar importance__heading
#> 1 X132 X134 1.299380      heading               y           -38.01651
#> 2 X132 X136 2.449381      heading     temperature          -102.98302
#> 3 X132 X153 2.091923      heading               x           -79.35320
#> 4 X134 X136 2.358040      heading     temperature           -90.27973
#> 5 X134 X153 2.074241      heading               x           -85.02712
#> 6 X136 X153 2.666099      heading               x          -116.24619
#>   importance__speed importance__temperature importance__x importance__y
#> 1         -21.91712              72.8121621      87.60660      93.95872
#> 2         -61.17293             356.8824838      48.38459      93.52145
#> 3         -40.92496             -14.2831545     427.73816     156.12855
#> 4         -59.26039             310.6812842      61.33615     108.96507
#> 5         -52.41491               0.9936944     592.07832     116.43104
#> 6         -65.45161             240.9053814     507.61536      56.69574
#>   psi_only_with__heading psi_only_with__speed psi_only_with__temperature
#> 1               1.075471             1.119086                   2.190984
#> 2               1.204174             1.397388                  10.327628
#> 3               1.373681             1.478797                   1.828918
#> 4               1.314354             1.426005                   8.954655
#> 5               1.263598             1.321072                   2.115805
#> 6               1.244071             1.385574                   8.429779
#>   psi_only_with__x psi_only_with__y psi_without__heading psi_without__speed
#> 1         2.391807         2.453403             1.569450           1.403873
#> 2         3.610898         4.644631             3.726621           2.895746
#> 3        10.657002         5.103284             3.033689           2.334915
#> 4         3.778736         4.839496             3.443185           2.823389
#> 5        13.797459         4.318944             3.027265           2.408284
#> 6        15.896024         4.126917             4.343310           3.130579
#>   psi_without__temperature psi_without__x psi_without__y
#> 1                 1.244877       1.253464       1.232522
#> 2                 1.586215       2.425775       2.353934
#> 3                 2.127711       1.709049       1.837195
#> 4                 1.628667       2.332406       2.270057
#> 5                 2.095194       1.516331       1.903884
#> 6                 2.007003       2.362495       2.615352