Cumulative Sum of Distances Between Consecutive Cases in a Time Series

Demonstration function to compute the sum of distances between consecutive cases in a time series.

psi_auto_distance(x = NULL, path = NULL, distance = "euclidean")

Arguments

x

(required, zoo object or matrix) univariate or multivariate time series with no NAs. Default: NULL

path

(optional, data frame) result of psi_cost_path_ignore_blocks(), only required if blocks in the least cost path were ignored. Default: NULL

distance

(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset distances. Default: "euclidean".

Value

numeric value

See also

Examples

#simulate a zoo object
x <- zoo_simulate(
  seed = 1
)

#sum of distances
psi_auto_distance(
  x = x,
  distance = "euclidean"
)
#> [1] 9.754384