Least Cost Path

Demonstration function to compute the least cost path within a least cost matrix.

Usage

psi_cost_path(dist_matrix = NULL, cost_matrix = NULL, diagonal = TRUE)

Arguments

dist_matrix

(required, numeric matrix) Distance matrix generated by psi_dist_matrix(). Default: NULL

cost_matrix

(required, numeric matrix) Cost matrix generated from the distance matrix with psi_cost_matrix(). Default: NULL

diagonal

(optional, logical) If TRUE, diagonals are used during the least-cost path search. Default: TRUE

Value

data frame

See also

Examples

#simulate two time series
tsl <- tsl_simulate(
  n = 2,
  seed = 1
)

if(interactive()){
  tsl_plot(tsl = tsl)
}

#distance matrix
dist_matrix <- psi_dist_matrix(
  x = tsl[[1]],
  y = tsl[[2]],
  distance = "euclidean"
)

#cost matrix
cost_matrix <- psi_cost_matrix(
  dist_matrix = dist_matrix,
  diagonal = FALSE
)

#diagonal least cost path
#------------------------
cost_path <- psi_cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix
)

if(interactive()){
  utils_matrix_plot(
    m = cost_matrix,
    path = cost_path
    )
}

#orthogonal least cost path
#--------------------------
cost_path <- psi_cost_path(
  dist_matrix = dist_matrix,
  cost_matrix = cost_matrix,
  diagonal = FALSE
)

if(interactive()){
  utils_matrix_plot(
    m = cost_matrix,
    path = cost_path
  )
}