Aggregation of Time Series List

Function to aggregate zoo objects within a time series list. This function supports progress bars generated by the progressr package. See examples.

This function also accepts a parallelization setup via future::plan(), but it might only be worth it for large time series lists.

Objective

Time series aggregation involves grouping observations and summarizing group values with a statistical function. This operation is useful to:

Decrease (downsampling) the temporal resolution of a time series.

Highlight particular states of a time series over time. For example, a daily temperature series can be aggregated by month using max to represent the highest temperatures each month.

Transform irregular time series into regular.

This function aggregates time series lists with overlapping times. Please check such overlap by assessing the columns "begin" and "end " of the data frame resulting from df <- tsl_time(tsl = tsl). Aggregation will be limited by the shortest time series in your time series list. To aggregate non-overlapping time series, please subset the individual components of tsl one by one either using tsl_subset() or the syntax tsl = my_tsl[[i]].

Methods

Any function returning a single number from a numeric vector can be used to aggregate a time series list. Quoted and unquoted function names can be used. Additional arguments to these functions can be passed via the argument .... Typical examples are:

mean or "mean": see mean().

median or "median": see stats::median().

quantile or "quantile": see stats::quantile().

min or "min": see min().

max or "max": see max().

sd or "sd": to compute standard deviation, see stats::sd().

var or "var": to compute the group variance, see stats::var().

length or "length": to compute group length.

sum or "sum": see sum().

f_slope or "f_slope": to compute the group slope, see f_slope().

tsl_aggregate(tsl = NULL, new_time = NULL, method = mean, ...)

Arguments

tsl

(required, list) Time series list. Default: NULL

new_time

(required, numeric, numeric vector, Date vector, POSIXct vector, or keyword) Definition of the aggregation pattern. The available options are:

numeric vector: only for the "numeric" time class, defines the breakpoints for time series aggregation.

"Date" or "POSIXct" vector: as above, but for the time classes "Date" and "POSIXct." In any case, the input vector is coerced to the time class of the tsl argument.

numeric: defines fixed time intervals in the units of tsl for time series aggregation. Used as is when the time class is "numeric", and coerced to integer and interpreted as days for the time classes "Date" and "POSIXct".

keyword (see utils_time_units()): the common options for the time classes "Date" and "POSIXct" are: "millennia", "centuries", "decades", "years", "quarters", "months", and "weeks". Exclusive keywords for the "POSIXct" time class are: "days", "hours", "minutes", and "seconds". The time class "numeric" accepts keywords coded as scientific numbers, from "1e8" to "1e-8".

method

(required, function name) Name of a standard or custom function to aggregate numeric vectors. Typical examples are mean, max,min, median, and quantile. Default: mean.

...

(optional) further arguments for method.

Value

time series list

See also

Examples

#parallelization setup (not worth it for this data size)
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
if(interactive()){
  progressr::handlers(global = TRUE)
}

# daily aggregation
#----------------------------------

#flight paths of five albatrosses
#scaled to improve visualization
tsl <- tsl_initialize(
  x = albatross,
  id_column = "id",
  time_column = "time"
) |>
  tsl_transform(
    f = f_scale
  )

#plot time series
if(interactive()){
  tsl_plot(
    tsl = tsl,
    guide_columns = 4
    )
}

#check time features
tsl_time(tsl)
#>      name rows   class units    length resolution               begin
#> X132 X132 1231 POSIXct  days  83.81359 0.06814113 2008-06-23 17:58:44
#> X134 X134 1051 POSIXct  days  70.12556 0.06678624 2008-05-31 14:59:31
#> X136 X136 1401 POSIXct  days  96.81472 0.06915337 2008-06-23 17:58:01
#> X153 X153 1300 POSIXct  days  84.62737 0.06514809 2008-07-10 23:57:31
#> X156 X156 1506 POSIXct  days 116.37640 0.07732651 2008-06-19 02:58:55
#>                      end     keywords
#> X132 2008-09-15 13:30:17 months, ....
#> X134 2008-08-09 18:00:18 months, ....
#> X136 2008-09-28 13:31:13 quarters....
#> X153 2008-10-03 15:00:56 months, ....
#> X156 2008-10-13 12:00:56 quarters....
#many observations per day
#different starting and ending dates

#aggregation: mean daily values
tsl_daily <- tsl_aggregate(
  tsl = tsl,
  new_time = "days",
  method = mean
)

#check time features
tsl_time(tsl_daily)
#>      name rows   class units length resolution               begin
#> X132 X132  135 POSIXct  days    134          1 2008-05-31 14:59:31
#> X134 X134  135 POSIXct  days    134          1 2008-05-31 14:59:31
#> X136 X136  135 POSIXct  days    134          1 2008-05-31 14:59:31
#> X153 X153  135 POSIXct  days    134          1 2008-05-31 14:59:31
#> X156 X156  135 POSIXct  days    134          1 2008-05-31 14:59:31
#>                      end     keywords
#> X132 2008-10-12 14:59:31 quarters....
#> X134 2008-10-12 14:59:31 quarters....
#> X136 2008-10-12 14:59:31 quarters....
#> X153 2008-10-12 14:59:31 quarters....
#> X156 2008-10-12 14:59:31 quarters....
#one observation per day
#same starting and ending dates

if(interactive()){
  tsl_plot(
    tsl = tsl_daily,
    guide_columns = 4
  )
}

# other supported keywords
#----------------------------------

#simulate full range of calendar dates
tsl <- tsl_simulate(
  n = 2,
  rows = 1000,
  time_range = c(
    "0000-01-01",
    as.character(Sys.Date())
    )
)

#mean value by millennia (extreme case!!!)
tsl_millennia <- tsl_aggregate(
  tsl = tsl,
  new_time = "millennia",
  method = mean
)

if(interactive()){
  tsl_plot(tsl_millennia)
}

#max value by centuries
tsl_centuries <- tsl_aggregate(
  tsl = tsl,
  new_time = "centuries",
  method = max
)

if(interactive()){
  tsl_plot(tsl_centuries)
}

#quantile 0.75 value by centuries
tsl_centuries <- tsl_aggregate(
  tsl = tsl,
  new_time = "centuries",
  method = stats::quantile,
  probs = 0.75 #argument of stats::quantile()
)

if(interactive()){
  tsl_plot(tsl_centuries)
}
#disable parallelization
future::plan(
  future::sequential
)