Renames Columns of Zoo Objects in a Time Series List

Renames Columns of Zoo Objects in a Time Series List

tsl_colnames_set(tsl = NULL, new_names = NULL)

Arguments

tsl

(required, list) Time series list. Default: NULL

new_names

(required, list) Named list. List names should match old column names in tsl, and each named item should contain a character string with the new name. For example, colnames = list(old_name = "new_name") changes the name of the column "old_name" to "new_name".

Value

time series list

See also

Examples

#generate example data
tsl <- tsl_simulate(cols = 3)

#list all column names
tsl_colnames_get(
  tsl = tsl,
  names = "all"
)
#> $A
#> [1] "a" "b" "c"
#> 
#> $B
#> [1] "a" "b" "c"
#> 

#rename columns
tsl <- tsl_colnames_set(
  tsl = tsl,
  new_names = list(
    a = "new_name_1",
    b = "new_name_1",
    c = "new_name_3"
  )
)

#check result
tsl_colnames_get(
  tsl = tsl,
  names = "all"
)
#> $A
#> [1] "new_name_1" "new_name_1" "new_name_3"
#> 
#> $B
#> [1] "new_name_1" "new_name_1" "new_name_3"
#>