Append Suffix to Column Names of Time Series List

Append Suffix to Column Names of Time Series List

Usage

tsl_colnames_suffix(tsl = NULL, suffix = NULL)

Arguments

tsl

(required, list) Time series list. Default: NULL

suffix

(optional, character string) String to append to the column names. Default: NULL.

Value

time series list

See also

Examples

tsl <- tsl_simulate()

tsl_colnames_get(tsl = tsl)
#> $A
#> [1] "a" "b" "c" "d" "e"
#> 
#> $B
#> [1] "a" "b" "c" "d" "e"
#> 

tsl <- tsl_colnames_suffix(
  tsl = tsl,
  suffix = "_my_suffix"
)

tsl_colnames_get(tsl = tsl)
#> $A
#> [1] "a_my_suffix" "b_my_suffix" "c_my_suffix" "d_my_suffix" "e_my_suffix"
#> 
#> $B
#> [1] "a_my_suffix" "b_my_suffix" "c_my_suffix" "d_my_suffix" "e_my_suffix"
#>