Validity Assessment of Time Series Lists

Validity assessment of time series lists. The required features for a valid time series list are:

Argument tsl is a list.

List tsl has unique names.

All elements in tsl are zoo objects.

All zoo objects have the attribute "name".

All zoo objects have unique values in the attribute "name".

Names of the list slots and the zoo objects are the same.

All zoo objects have at least one shared column name.

All columns in the zoo objects are numeric.

All zoo objects have zero NA cases.

This function analyzes a time series list and returns it without modification, but issues a warning if any condition is not met, and an explanation on how to solve the issue.

Usage

tsl_diagnose(tsl = NULL, full = TRUE)

Arguments

tsl

(required, list) Time series list. Default: NULL

full

(optional, logical) If TRUE, the function runs checks looking at all the individual values of the time series list. This makes the test slower, but more comprehensive. Default: TRUE

Value

time series list

See also

Examples

#simulate time series list with NA data
tsl <- tsl_simulate(
  n = 3,
  na_fraction = 0.1
)

#adding a few additional issues:

#set names of one time series to uppercase
colnames(tsl[[2]]) <- toupper(colnames(tsl[[2]]))

#change name of one column in one time series
colnames(tsl[[3]])[2] <- "x"

#remove list names
names(tsl) <- NULL

#diagnose issues in tsl
tsl <- tsl_diagnose(
  tsl = tsl,
  full = TRUE
)
#> This time series list is NOT VALID.
#> The issue/s to address are:
#>   - elements of 'tsl' must be named: use tsl_names_set() or names(tsl) <- c(...) to fix this issue.
#>   - zoo objects in 'tsl' must have at least one shared column: use tsl_colnames_get() to identify shared and/or exclusive columns.
#>   - zoo objects in 'tsl' have NA cases: interpolate or remove them with tsl_handle_NA() to fix this issue.

#repair time series list
tsl <- tsl_repair(
  tsl = tsl,
  full = TRUE
)
#> Naming items in object 'tsl'
#> Imputing NA values in zoo objects.

#diagnose again
tsl <- tsl_diagnose(
  tsl = tsl,
  full = TRUE
)