Coerce Elements of Time Series List to same Time Class

Coerce Elements of Time Series List to same Time Class

Usage

tsl_time_class_set(tsl = NULL, to = NULL)

Arguments

tsl

(required, list) Time series list. Default: NULL

to

(required, class name) class to coerce x to. Either "Date", "POSIXct", "integer" or "numeric". If NULL, coerces to the majority time class. Default: NULL

Value

time series list

See also

Examples

#TODO missing example