Check Input Arguments

Check Input Arguments

utils_check_distantia_args(
  tsl = NULL,
  distance = "euclidean",
  diagonal = TRUE,
  weighted = TRUE,
  ignore_blocks = FALSE,
  lock_step = FALSE,
  repetitions = 0,
  permutation = "restricted_by_row",
  block_size = 3,
  seed = 1,
  robust = TRUE
)

Arguments

tsl

(required, list) Time series list. Default: NULL

distance

(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset distances. Default: "euclidean".

diagonal

(optional, logical vector). If TRUE, diagonals are included in the computation of the cost matrix. Default: FALSE.

weighted

(optional, logical vector) If TRUE, diagonal is set to TRUE, and diagonal cost is weighted by a factor of 1.414214. Default: FALSE.

ignore_blocks

(optional, logical vector). If TRUE, blocks of consecutive path coordinates are trimmed to avoid inflating the psi distance. Default: FALSE.

lock_step

(optional, logical vector) If TRUE, time series are compared row wise and no least-cost path is computed. Default: FALSE.

repetitions

(optional, integer vector) number of permutations to compute the p-value (interpreted as the probability of finding a smaller dissimilarity on permuted versions of the sequences) of the psi distance. If 0, p-values are not computed. Otherwise, the minimum is 2. Default: 0

permutation

(optional, character vector) permutation method. Valid values are listed below from higher to lower randomness:

"free": unrestricted shuffling of rows and columns. Ignores block_size.

"free_by_row": unrestricted shuffling of complete rows. Ignores block size.

"restricted": restricted shuffling of rows and columns within blocks.

"restricted_by_row": restricted shuffling of rows within blocks.

block_size

(optional, integer vector) vector with block sizes in rows for the restricted permutation test. A block of size 3 indicates that a row can only be permuted within a block of 3 adjacent rows. If several values are provided, one is selected at random separately for each time series on each repetition. Only relevant when permutation methods are "restricted" or "restricted_by_row". Default: 3.

seed

(optional, integer) initial random seed to use for replicability when computing p-values. Default: 1

robust

(required, logical). If TRUE, importance scores are computed using the least cost path used to compute the psi dissimilarity between the two full time series. Setting it to FALSE allows to replicate importance scores of the previous versions of this package. Default: TRUE

Value

list

See also