Data Frame with Contribution of Individual Variables to Dissimilarity to Wide Format

Takes as input a data frame returned by distantia_importance() to return a data frame with one pair of time series per row, and the following columns:

most_similar: name of the variable with the highest contribution to similarity (most negative value in the importance column) for each pair of time series.

most_dissimilar: name of the variable with the highest contribution to dissimilarity (most positive value in the importance column) for each pair of time series.

importance__variable_name: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.

psi_only_with__variable_name: dissimilarity of the two time series when only using the given variable.

psi_without__variable_name: dissimilarity of the two time series when removing the given variable.

Usage

utils_importance_df_to_wide(df = NULL, sep = "__")

Arguments

df

(required, data frame) Output of distantia_importance(). Default: NULL

sep

(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".

Value

data frame

See also

Examples

#prepare time series
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  name_column = "name",
  time_column = "time"
) |>
  tsl_transform(
    f = f_scale_global
  )

#importance data frame
df <- distantia_importance(
  tsl = tsl
)
df
#>         x      y       psi    variable importance               effect
#> 1 Germany  Spain 1.0360268         evi  191.78590 decreases similarity
#> 2 Germany  Spain 1.0360268    rainfall -176.40303 increases similarity
#> 3 Germany  Spain 1.0360268 temperature  176.27550 decreases similarity
#> 4 Germany Sweden 0.7285701         evi   91.17026 decreases similarity
#> 5 Germany Sweden 0.7285701    rainfall  -53.79797 increases similarity
#> 6 Germany Sweden 0.7285701 temperature   53.62633 decreases similarity
#> 7   Spain Sweden 1.0237508         evi  262.88414 decreases similarity
#> 8   Spain Sweden 1.0237508    rainfall -266.25958 increases similarity
#> 9   Spain Sweden 1.0237508 temperature  266.29320 decreases similarity
#>   psi_difference psi_without psi_only_with  distance diagonal lock_step robust
#> 1      1.9869533   1.0359739     3.0229272 euclidean     TRUE     FALSE   TRUE
#> 2     -1.8275826   2.8033119     0.9757293 euclidean     TRUE     FALSE   TRUE
#> 3      1.8262614   0.9758464     2.8021078 euclidean     TRUE     FALSE   TRUE
#> 4      0.6642392   0.7285448     1.3927840 euclidean     TRUE     FALSE   TRUE
#> 5     -0.3919559   1.1048296     0.7128736 euclidean     TRUE     FALSE   TRUE
#> 6      0.3907054   0.7129364     1.1036418 euclidean     TRUE     FALSE   TRUE
#> 7      2.6912784   1.0236697     3.7149482 euclidean     TRUE     FALSE   TRUE
#> 8     -2.7258345   3.6539963     0.9281619 euclidean     TRUE     FALSE   TRUE
#> 9      2.7261787   0.9284271     3.6546058 euclidean     TRUE     FALSE   TRUE

#to wide format
df_wide <- utils_importance_df_to_wide(
  df = df
)
df_wide
#>         x      y       psi most_similar most_dissimilar importance__evi
#> 1 Germany  Spain 1.0360268     rainfall             evi       191.78590
#> 2 Germany Sweden 0.7285701     rainfall             evi        91.17026
#> 3   Spain Sweden 1.0237508     rainfall     temperature       262.88414
#>   importance__rainfall importance__temperature psi_only_with__evi
#> 1           -176.40303               176.27550           3.022927
#> 2            -53.79797                53.62633           1.392784
#> 3           -266.25958               266.29320           3.714948
#>   psi_only_with__rainfall psi_only_with__temperature psi_without__evi
#> 1               0.9757293                   2.802108        1.0359739
#> 2               0.7128736                   1.103642        0.7285448
#> 3               0.9281619                   3.654606        1.0236697
#>   psi_without__rainfall psi_without__temperature
#> 1              2.803312                0.9758464
#> 2              1.104830                0.7129364
#> 3              3.653996                0.9284271