Handles Line Colors for Sequence Plots

This is an internal function, but can be used to better understand how line colors are handled within other plotting functions.

Usage

utils_line_color(x = NULL, line_color = NULL)

Arguments

x

(required, sequence) zoo object or time series list. Default: NULL

line_color

(optional, character vector) vector of colors for the time series columns. Selected palette depends on the number of columns to plot. Default: NULL

Value

color vector

See also