Handles Time Column in a List of Data Frames

Handles Time Column in a List of Data Frames

Usage

utils_prepare_time(x = NULL, time_column = NULL, lock_step = FALSE)

Arguments

x

(required, named list of data frames). List with named data frames. Default: NULL.

time_column

(optional if lock_step = FALSE, and required otherwise, column name) Name of numeric column representing time. Default: NULL.

lock_step

(optional, logical) If TRUE, all input sequences are subset to their common times according to the values in the time_column. Default: FALSE.

Value

List of data frames

See also