Get Time Features from Zoo Objects

This function generates a data frame summarizing the time features (class, length, resolution, and others) of zoo time series.

Usage

zoo_time(x = NULL, keywords = c("resample", "aggregate"))

Arguments

x

(required, zoo object) Zoo time series to analyze. Default: NULL.

keywords

(optional, character string or vector) Defines what keywords are returned. If "aggregate", returns valid keywords for zoo_aggregate(). If "resample", returns valid keywords for zoo_resample(). If both, returns all valid keywords. Default: c("aggregate", "resample").

Value

Data frame with the following columns:

name (string): time series name.

rows (integer): number of observations.

class (string): time class, one of "Date", "POSIXct", or "numeric."

units (string): units of the time series.

length (numeric): total length of the time series expressed in units.

resolution (numeric): average interval between observations expressed in units.

begin (date or numeric): begin time of the time series.

end (date or numeric): end time of the time series.

keywords (character vector): valid keywords for tsl_aggregate() or tsl_resample(), depending on the value of the argument keywords.

See also

Examples

#simulate a zoo time series
x <- zoo_simulate(
  rows = 150,
  time_range = c(
    Sys.Date() - 365,
    Sys.Date()
  ),
  irregular = TRUE
)

#time data frame
zoo_time(
  x = x
)
#>   name rows class units length resolution      begin        end     keywords
#> 1    A  150  Date  days    365   2.449664 2023-12-10 2024-12-09 years, q....