Aggregate Dissimilarity Analysis Data Frames Across Parameter Combinations

The functions distantia() and distantia_importance() allow dissimilarity assessments based on different combinations of arguments at once. For example, when the argument distance is set to c("euclidean", "manhattan"), the output data frame will show two dissimilarity scores for each pair of compared time series, one based on euclidean distances, and another based on manhattan distances.

When df is the result of distantia(), the input data is grouped by pairs of time series, and the function f is applied to the column "psi" by group

When df is the result of distantia_importance(), the input data is grouped by pairs of time series and variables, and the function f is applied to the columns "importance", "psi_only_with" and "psi_without" by group. However, if the values TRUE and FALSE appear in the column "robust" (which is not allowed by default in distantia_importance()), then the aggregation is cancelled with an error, as the results of both methods should not be aggregated together.

If psi scores smaller than zero occur in the aggregated output, then the the smaller psi value is added to the column psi to start dissimilarity scores at zero.

If there are no different combinations of arguments in the input data frame, no aggregation happens, but all parameter columns are removed.

Usage

distantia_aggregate(df = NULL, f = mean, ...)

Arguments

df

(required, data frame) Output of distantia() or distantia_importance(). Default: NULL

f

(optional, function) Function to summarize psi scores (for example, mean) when there are several combinations of parameters in df. Ignored when there is a single combination of arguments in the input. Default: mean

...

(optional, arguments of f) Further arguments to pass to the function f.

Value

data frame

See also

Examples

#three time series
#climate and ndvi in Fagus sylvatica stands in Spain, Germany, and Sweden
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  id_column = "site",
  time_column = "date"
) |>
  tsl_transform(
    f = f_scale
  )

if(interactive()){
  tsl_plot(
    tsl = tsl,
    guide_columns = 3
    )
}

#distantia with multiple parameter combinations
#-------------------------------------
df_multiple <- distantia(
  tsl = tsl,
  distance = c("euclidean", "manhattan"),
  lock_step = c(TRUE, FALSE)
)
#> Loading required package: foreach
#> Loading required package: future

df_multiple[, c(
  "x",
  "y",
  "distance",
  "lock_step",
  "psi"
)]
#>          x      y  distance lock_step       psi
#> 1  Germany  Spain euclidean      TRUE 1.3061327
#> 2  Germany Sweden euclidean      TRUE 0.8576700
#> 3    Spain Sweden euclidean      TRUE 1.4708497
#> 4  Germany  Spain manhattan      TRUE 1.2698922
#> 5  Germany Sweden manhattan      TRUE 0.8591195
#> 6    Spain Sweden manhattan      TRUE 1.4890286
#> 7  Germany  Spain euclidean     FALSE 1.3429956
#> 8  Germany Sweden euclidean     FALSE 0.8571217
#> 9    Spain Sweden euclidean     FALSE 1.4803954
#> 10 Germany  Spain manhattan     FALSE 1.2850779
#> 11 Germany Sweden manhattan     FALSE 0.8527760
#> 12   Spain Sweden manhattan     FALSE 1.4959250

#aggregation using means
df <- distantia_aggregate(
  df = df_multiple,
  f = mean
)

df
#>         x      y       psi
#> 1 Germany  Spain 1.3010246
#> 2 Germany Sweden 0.8566718
#> 3   Spain Sweden 1.4840497

#importance with multiple parameter combinations
#-------------------------------------
df_multiple <- distantia_importance(
  tsl = tsl,
  distance = c("euclidean", "manhattan"),
  lock_step = c(FALSE, TRUE)
)

df_multiple[, c(
  "x",
  "y",
  "variable",
  "distance",
  "lock_step",
  "importance"
)]
#>          x      y    variable  distance lock_step importance
#> 1  Germany  Spain         evi euclidean     FALSE   6.512321
#> 2  Germany  Spain    rainfall euclidean     FALSE  12.505764
#> 3  Germany  Spain temperature euclidean     FALSE -26.509115
#> 4  Germany Sweden         evi euclidean     FALSE  29.026504
#> 5  Germany Sweden    rainfall euclidean     FALSE  -4.209397
#> 6  Germany Sweden temperature euclidean     FALSE -26.661768
#> 7    Spain Sweden         evi euclidean     FALSE  -6.625437
#> 8    Spain Sweden    rainfall euclidean     FALSE  -4.416941
#> 9    Spain Sweden temperature euclidean     FALSE  13.668290
#> 10 Germany  Spain         evi manhattan     FALSE   5.692418
#> 11 Germany  Spain    rainfall manhattan     FALSE  19.828337
#> 12 Germany  Spain temperature manhattan     FALSE -31.724040
#> 13 Germany Sweden         evi manhattan     FALSE  31.231755
#> 14 Germany Sweden    rainfall manhattan     FALSE  -1.117102
#> 15 Germany Sweden temperature manhattan     FALSE -31.510126
#> 16   Spain Sweden         evi manhattan     FALSE -17.569115
#> 17   Spain Sweden    rainfall manhattan     FALSE   5.323929
#> 18   Spain Sweden temperature manhattan     FALSE  11.959892
#> 19 Germany  Spain         evi euclidean      TRUE   0.241829
#> 20 Germany  Spain    rainfall euclidean      TRUE  19.051052
#> 21 Germany  Spain temperature euclidean      TRUE -30.814944
#> 22 Germany Sweden         evi euclidean      TRUE  28.539736
#> 23 Germany Sweden    rainfall euclidean      TRUE  -4.845232
#> 24 Germany Sweden temperature euclidean      TRUE -25.011608
#> 25   Spain Sweden         evi euclidean      TRUE -22.912397
#> 26   Spain Sweden    rainfall euclidean      TRUE   9.732110
#> 27   Spain Sweden temperature euclidean      TRUE  12.501949
#> 28 Germany  Spain         evi manhattan      TRUE   1.900935
#> 29 Germany  Spain    rainfall manhattan      TRUE  21.154702
#> 30 Germany  Spain temperature manhattan      TRUE -29.727709
#> 31 Germany Sweden         evi manhattan      TRUE  29.249949
#> 32 Germany Sweden    rainfall manhattan      TRUE  -3.444891
#> 33 Germany Sweden temperature manhattan      TRUE -26.684626
#> 34   Spain Sweden         evi manhattan      TRUE -25.080233
#> 35   Spain Sweden    rainfall manhattan      TRUE  10.764640
#> 36   Spain Sweden temperature manhattan      TRUE  12.742187

#aggregation using means
df <- distantia_aggregate(
  df = df_multiple,
  f = mean
)

df
#>         x      y    variable       psi importance psi_without psi_only_with
#> 1 Germany  Spain         evi 1.3010246   3.586876   1.2811692     1.3281469
#> 2 Germany Sweden         evi 0.8566718  29.511986   0.7807204     1.0335199
#> 3   Spain Sweden         evi 1.4840497 -18.046795   1.5579973     1.2901568
#> 4 Germany  Spain    rainfall 1.3010246  18.134964   1.1741316     1.4091905
#> 5 Germany Sweden    rainfall 0.8566718  -3.404155   0.8726797     0.8434902
#> 6   Spain Sweden    rainfall 1.4840497   5.350935   1.4606764     1.5400981
#> 7 Germany  Spain temperature 1.3010246 -29.693952   1.3907864     1.0048642
#> 8 Germany Sweden temperature 0.8566718 -27.467032   0.9190628     0.6838117
#> 9   Spain Sweden temperature 1.4840497  12.718079   1.4467168     1.6354357

df[, c(
  "x",
  "y",
  "variable",
  "importance"
)]
#>         x      y    variable importance
#> 1 Germany  Spain         evi   3.586876
#> 2 Germany Sweden         evi  29.511986
#> 3   Spain Sweden         evi -18.046795
#> 4 Germany  Spain    rainfall  18.134964
#> 5 Germany Sweden    rainfall  -3.404155
#> 6   Spain Sweden    rainfall   5.350935
#> 7 Germany  Spain temperature -29.693952
#> 8 Germany Sweden temperature -27.467032
#> 9   Spain Sweden temperature  12.718079

#distantia with a single parameter combination
#------------------------------
df_multiple <- distantia(
  tsl = tsl,
  distance = "euclidean",
  lock_step = TRUE
)

df_multiple
#>         x      y  distance diagonal weighted ignore_blocks lock_step      psi
#> 1 Germany  Spain euclidean     TRUE     TRUE         FALSE      TRUE 1.306133
#> 2 Germany Sweden euclidean     TRUE     TRUE         FALSE      TRUE 0.857670
#> 3   Spain Sweden euclidean     TRUE     TRUE         FALSE      TRUE 1.470850

#no aggregation happens, but all parameter columns are removed
df <- distantia_aggregate(
  df = df_multiple,
  f = mean
)

df
#>         x      y      psi
#> 1 Germany  Spain 1.306133
#> 2 Germany Sweden 0.857670
#> 3   Spain Sweden 1.470850