Hierarchical Clustering of Dissimilarity Analysis Data Frames

This function combines the dissimilarity scores computed by distantia(), the agglomerative clustering methods provided by stats::hclust(), and the clustering optimization method implemented in utils_cluster_hclust_optimizer() to help group together time series with similar features.

When clusters = NULL, the function utils_cluster_hclust_optimizer() is run underneath to perform a parallelized grid search to find the number of clusters maximizing the overall silhouette width of the clustering solution (see utils_cluster_silhouette()). When method = NULL as well, the optimization also includes all methods available in stats::hclust() in the grid search. The grid search function supports parallelization via future::plan() and a progress bar generated by the progressr package (see Examples).

Usage

distantia_cluster_hclust(df = NULL, clusters = NULL, method = "complete")

Arguments

df

(required, data frame) Output of distantia(). Default: NULL

clusters

(required, integer) Number of groups to generate. If NULL (default), utils_cluster_kmeans_optimizer() is used to find the number of clusters that maximizes the mean silhouette width of the clustering solution (see utils_cluster_silhouette()). Default: NULL

method

(optional, character string) Argument of stats::hclust() defining the agglomerative method. One of: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Unambiguous abbreviations are accepted as well. If NULL (default), utils_cluster_hclust_optimizer() finds the optimal method. Default: "complete".

Value

list:

cluster_object: hclust object for further analyses and custom plotting.

clusters: integer, number of clusters.

silhouette_width: mean silhouette width of the clustering solution.

df: data frame with time series names, their cluster label, and their individual silhouette width scores.

d: psi distance matrix used for clustering.

optimization: only if clusters = NULL, data frame with optimization results from utils_cluster_hclust_optimizer().

See also

Examples

#for large datasets, parallelization accelerates cluster optimization
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
# progressr::handlers(global = TRUE)

#daily covid prevalence in California counties
data("covid_prevalence")

#load as tsl
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
)
#> Structural issues:
#> ------------------
#> 
#>   - time in all time series must be of the same class: use lapply(tsl, function(x) class(zoo::index(x))) to identify and remove or modify the objects with a mismatching class.

#subset 10 elements to accelerate example execution
tsl <- tsl_subset(
  tsl = tsl,
  names = 1:10
)

#aggregateto monthly data to accelerate example execution
tsl <- tsl_aggregate(
  tsl = tsl,
  new_time = "months",
  fun = sum
)
#> Structural issues:
#> ------------------
#> 
#>   - time in all time series must be of the same class: use lapply(tsl, function(x) class(zoo::index(x))) to identify and remove or modify the objects with a mismatching class.
#> Structural issues:
#> ------------------
#> 
#>   - time in all time series must be of the same class: use lapply(tsl, function(x) class(zoo::index(x))) to identify and remove or modify the objects with a mismatching class.
#> Renaming zoo time series from 'Alameda' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Butte' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Contra_Costa' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'El_Dorado' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Fresno' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Humboldt' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Imperial' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Kern' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Kings' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().
#> Renaming zoo time series from 'Los_Angeles' to ''.
#> Structural issues:
#> ------------------
#> 
#>   - list and time series names must match and be unique: reset names with distantia::tsl_names_set().

if(interactive()){
  #plotting first three time series
  tsl_plot(
    tsl = tsl_subset(
      tsl = tsl,
      names = 1:3
    ),
    guide_columns = 3
  )
}

#dissimilarity analysis
distantia_df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)
#> Structural issues:
#> ------------------
#> 
#>   - time in all time series must be of the same class: use lapply(tsl, function(x) class(zoo::index(x))) to identify and remove or modify the objects with a mismatching class.

#hierarchical clustering with a given number of clusters
#-------------------------------------------------------
distantia_clust <- distantia_cluster_hclust(
  df = distantia_df,
  clusters = 5, #arbitrary number!
  method = "complete"
)

#names of the output object
names(distantia_clust)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"               

#cluster object
distantia_clust$cluster_object
#> 
#> Call:
#> stats::hclust(d = d_dist, method = method)
#> 
#> Cluster method   : complete 
#> Number of objects: 10 
#> 

#distance matrix used for clustering
distantia_clust$d
#>                Alameda     Butte Contra_Costa El_Dorado    Fresno  Humboldt
#> Butte        0.6287502                                                     
#> Contra_Costa 0.1417180 0.5537658                                           
#> El_Dorado    0.6043963 0.2791084    0.5562701                              
#> Fresno       0.5396143 0.5514503    0.4892774 0.6451476                    
#> Humboldt     0.7198591 0.4681926    0.6762226 0.3971290 0.8496143          
#> Imperial     0.7524957 0.9603572    0.7399531 0.9982246 0.4790492 1.1988480
#> Kern         0.6607005 0.4648086    0.5876087 0.5807203 0.2627634 0.8162986
#> Kings        0.8711298 0.9578573    0.8231969 1.0517096 0.4505311 1.2177286
#> Los_Angeles  0.5425064 0.8740719    0.5263465 0.8598873 0.4542238 1.0282498
#>               Imperial      Kern     Kings
#> Butte                                     
#> Contra_Costa                              
#> El_Dorado                                 
#> Fresno                                    
#> Humboldt                                  
#> Imperial                                  
#> Kern         0.6753489                    
#> Kings        0.4748470 0.6421878          
#> Los_Angeles  0.4960491 0.5922560 0.4632569

#number of clusters
distantia_clust$clusters
#> [1] 5

#clustering data frame
#group label in column "cluster"
#negatives in column "silhouette_width" higlight anomalous cluster assignation
distantia_clust$df
#>            name cluster silhouette_width
#> 1       Alameda       1       0.76386522
#> 2         Butte       2       0.26465487
#> 3  Contra_Costa       1       0.73680033
#> 4     El_Dorado       2       0.41737146
#> 5        Fresno       3       0.41915007
#> 6      Humboldt       2       0.38017839
#> 7      Imperial       4       0.00000000
#> 8          Kern       3       0.57428048
#> 9         Kings       5       0.02440797
#> 10  Los_Angeles       5       0.06610664

#mean silhouette width of the clustering solution
distantia_clust$silhouette_width
#> [1] 0.3646815

#plot
if(interactive()){

  clust <- distantia_clust$cluster_object
  k <- distantia_clust$clusters

  #tree plot
  plot(
    x = clust,
    hang = -1
  )

  #highlight groups
  stats::rect.hclust(
    tree = clust,
    k = k,
    cluster = stats::cutree(
      tree = clust,
      k = k
    )
  )

}

#optimized hierarchical clustering
#---------------------------------

#auto-optimization of clusters and method
distantia_clust <- distantia_cluster_hclust(
  df = distantia_df,
  clusters = NULL,
  method = NULL
)

#names of the output object
#a new object named "optimization" should appear
names(distantia_clust)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"                "optimization"    

#first rows of the optimization data frame
#optimized clustering in first row
head(distantia_clust$optimization)
#>   clusters   method silhouette_mean
#> 1        4   ward.D       0.3953577
#> 2        4  ward.D2       0.3953577
#> 3        4 complete       0.3953577
#> 4        4  average       0.3953577
#> 5        4 mcquitty       0.3953577
#> 6        3   ward.D       0.3784943

#plot
if(interactive()){

  clust <- distantia_clust$cluster_object
  k <- distantia_clust$clusters

  #tree plot
  plot(
    x = clust,
    hang = -1
  )

  #highlight groups
  stats::rect.hclust(
    tree = clust,
    k = k,
    cluster = stats::cutree(
      tree = clust,
      k = k
    )
  )

}

#disable parallelization
future::plan(
  future::sequential
)