K-means Clustering from Dissimilarity Scores

This function combines the dissimilarity scores computed by distantia(), the K-means clustering method implemented in stats::kmeans(), and the clustering optimization method implemented in utils_cluster_hclust_optimizer() to help group together time series with similar features.

When clusters = NULL, the function utils_cluster_hclust_optimizer() is run underneath to perform a parallelized grid search to find the number of clusters maximizing the overall silhouette width of the clustering solution (see utils_cluster_silhouette()). The grid search function supports parallelization via future::plan() and a progress bar generated by the progressr package (see Examples).

distantia_cluster_kmeans(df = NULL, clusters = NULL, seed = 1)

Arguments

df

(required, data frame) Output of distantia(). Default: NULL

clusters

(required, integer) Number of groups to generate. If NULL (default), utils_cluster_kmeans_optimizer() is used to find the number of clusters that maximizes the mean silhouette width of the clustering solution (see utils_cluster_silhouette()). Default: NULL

seed

(optional, integer) Random seed to be used during the K-means computation. Default: 1

Value

list:

cluster_object: kmeans object object for further analyses and custom plotting.

clusters: integer, number of clusters.

silhouette_width: mean silhouette width of the clustering solution.

df: data frame with time series names, their cluster label, and their individual silhouette width scores.

d: psi distance matrix used for clustering.

optimization: only if clusters = NULL, data frame with optimization results from utils_cluster_hclust_optimizer().

See also

Examples

#for large datasets, parallelization accelerates cluster optimization
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
if(interactive()){
  progressr::handlers(global = TRUE)
}

#daily covid prevalence in California counties
data("covid_prevalence")

#load as tsl and aggregate to monthly data to accelerate example execution
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
)

#subset 10 elements to accelerate example execution
tsl <- tsl_subset(
  tsl = tsl,
  names = 1:10
)

#aggregateto monthly data to accelerate example execution
tsl <- tsl_aggregate(
  tsl = tsl,
  new_time = "months",
  fun = sum
)

if(interactive()){
  #plotting first three time series
  tsl_plot(
    tsl = tsl[1:3],
    guide_columns = 3
    )
}

#dissimilarity analysis
distantia_df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#kmeans with a given number of clusters
#-------------------------------------------------------
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = 5 #arbitrary number!
)

#names of the output object
names(distantia_kmeans)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"               

#kmeans object
distantia_kmeans$cluster_object
#> K-means clustering with 5 clusters of sizes 2, 2, 2, 1, 3
#> 
#> Cluster means:
#>      Alameda     Butte Contra_Costa El_Dorado    Fresno  Humboldt  Imperial
#> 1 0.07085902 0.5912580   0.07085902 0.5803332 0.5144459 0.6980408 0.7462244
#> 2 0.60015741 0.5081294   0.53844306 0.6129340 0.1313817 0.8329565 0.5771991
#> 3 0.61657321 0.1395542   0.55501793 0.1395542 0.5982990 0.4326608 0.9792909
#> 4 0.71985907 0.4681926   0.67622262 0.3971290 0.8496143 0.0000000 1.1988480
#> 5 0.72204396 0.9307621   0.69649882 0.9699405 0.4612680 1.1482755 0.3236320
#>        Kern     Kings Los_Angeles
#> 1 0.6241546 0.8471634   0.5344264
#> 2 0.1313817 0.5463595   0.5232399
#> 3 0.5227645 1.0047835   0.8669796
#> 4 0.8162986 1.2177286   1.0282498
#> 5 0.6365976 0.3127013   0.3197687
#> 
#> Clustering vector:
#>      Alameda        Butte Contra_Costa    El_Dorado       Fresno     Humboldt 
#>            1            3            1            3            2            4 
#>     Imperial         Kern        Kings  Los_Angeles 
#>            5            2            5            5 
#> 
#> Within cluster sum of squares by cluster:
#> [1] 0.03030158 0.13475326 0.09705527 0.00000000 0.61045676
#>  (between_SS / total_SS =  89.4 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      

#distance matrix used for clustering
distantia_kmeans$d
#>                Alameda     Butte Contra_Costa El_Dorado    Fresno  Humboldt
#> Alameda      0.0000000 0.6287502    0.1417180 0.6043963 0.5396143 0.7198591
#> Butte        0.6287502 0.0000000    0.5537658 0.2791084 0.5514503 0.4681926
#> Contra_Costa 0.1417180 0.5537658    0.0000000 0.5562701 0.4892774 0.6762226
#> El_Dorado    0.6043963 0.2791084    0.5562701 0.0000000 0.6451476 0.3971290
#> Fresno       0.5396143 0.5514503    0.4892774 0.6451476 0.0000000 0.8496143
#> Humboldt     0.7198591 0.4681926    0.6762226 0.3971290 0.8496143 0.0000000
#> Imperial     0.7524957 0.9603572    0.7399531 0.9982246 0.4790492 1.1988480
#> Kern         0.6607005 0.4648086    0.5876087 0.5807203 0.2627634 0.8162986
#> Kings        0.8711298 0.9578573    0.8231969 1.0517096 0.4505311 1.2177286
#> Los_Angeles  0.5425064 0.8740719    0.5263465 0.8598873 0.4542238 1.0282498
#>               Imperial      Kern     Kings Los_Angeles
#> Alameda      0.7524957 0.6607005 0.8711298   0.5425064
#> Butte        0.9603572 0.4648086 0.9578573   0.8740719
#> Contra_Costa 0.7399531 0.5876087 0.8231969   0.5263465
#> El_Dorado    0.9982246 0.5807203 1.0517096   0.8598873
#> Fresno       0.4790492 0.2627634 0.4505311   0.4542238
#> Humboldt     1.1988480 0.8162986 1.2177286   1.0282498
#> Imperial     0.0000000 0.6753489 0.4748470   0.4960491
#> Kern         0.6753489 0.0000000 0.6421878   0.5922560
#> Kings        0.4748470 0.6421878 0.0000000   0.4632569
#> Los_Angeles  0.4960491 0.5922560 0.4632569   0.0000000
#> attr(,"distantia_args")
#>    distance diagonal weighted ignore_blocks lock_step group
#> 1 euclidean     TRUE     TRUE         FALSE      TRUE     1
#> attr(,"type")
#> [1] "distantia_matrix"
#> attr(,"distance")
#> [1] "psi"

#number of clusters
distantia_kmeans$clusters
#> [1] 5

#clustering data frame
#group label in column "cluster"
distantia_kmeans$df
#>            name cluster silhouette_width
#> 1       Alameda       1        0.7638652
#> 2         Butte       3        0.4038600
#> 3  Contra_Costa       1        0.7368003
#> 4     El_Dorado       3        0.2971845
#> 5        Fresno       2        0.4303455
#> 6      Humboldt       4        0.0000000
#> 7      Imperial       5        0.1589591
#> 8          Kern       2        0.4973579
#> 9         Kings       5        0.1414957
#> 10  Los_Angeles       5        0.0833020

#mean silhouette width of the clustering solution
distantia_kmeans$silhouette_width
#> [1] 0.351317

#kmeans plot
# factoextra::fviz_cluster(
#   object = distantia_kmeans$cluster_object,
#   data = distantia_kmeans$d,
#   repel = TRUE
# )

#optimized kmeans
#---------------------------------
#auto-optimization of clusters and method
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = NULL
)

#names of the output object
#a new object named "optimization" should appear
names(distantia_kmeans)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"                "optimization"    

#first rows of the optimization data frame
#optimized clustering in first row
head(distantia_kmeans$optimization)
#>   clusters silhouette_mean
#> 1        4       0.3953577
#> 2        5       0.3513170
#> 3        2       0.3435351
#> 4        6       0.3107771
#> 5        7       0.3084695
#> 6        3       0.2760140

#disable parallelization
future::plan(
  future::sequential
)