K-Means Clustering of Dissimilarity Analysis Data Frames

This function combines the dissimilarity scores computed by distantia(), the K-means clustering method implemented in stats::kmeans(), and the clustering optimization method implemented in utils_cluster_hclust_optimizer() to help group together time series with similar features.

When clusters = NULL, the function utils_cluster_hclust_optimizer() is run underneath to perform a parallelized grid search to find the number of clusters maximizing the overall silhouette width of the clustering solution (see utils_cluster_silhouette()). The grid search function supports parallelization via future::plan() and a progress bar generated by the progressr package (see Examples).

Usage

distantia_cluster_kmeans(df = NULL, clusters = NULL, seed = 1)

Arguments

df

(required, data frame) Output of distantia(). Default: NULL

clusters

(required, integer) Number of groups to generate. If NULL (default), utils_cluster_kmeans_optimizer() is used to find the number of clusters that maximizes the mean silhouette width of the clustering solution (see utils_cluster_silhouette()). Default: NULL

seed

(optional, integer) Random seed to be used during the K-means computation. Default: 1

Value

list:

cluster_object: kmeans object object for further analyses and custom plotting.

clusters: integer, number of clusters.

silhouette_width: mean silhouette width of the clustering solution.

df: data frame with time series names, their cluster label, and their individual silhouette width scores.

d: psi distance matrix used for clustering.

optimization: only if clusters = NULL, data frame with optimization results from utils_cluster_hclust_optimizer().

See also

Examples

#for large datasets, parallelization accelerates cluster optimization
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
if(interactive()){
  progressr::handlers(global = TRUE)
}

#daily covid prevalence in California counties
data("covid_prevalence")

#load as tsl and aggregate to monthly data to accelerate example execution
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "county",
  time_column = "date"
)

#subset 10 elements to accelerate example execution
tsl <- tsl_subset(
  tsl = tsl,
  names = 1:10
)

#aggregateto monthly data to accelerate example execution
tsl <- tsl_aggregate(
  tsl = tsl,
  new_time = "months",
  fun = sum
)

if(interactive()){
  #plotting first three time series
  tsl_plot(
    tsl = tsl[1:3],
    guide_columns = 3
    )
}

#dissimilarity analysis
distantia_df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#kmeans with a given number of clusters
#-------------------------------------------------------
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = 5 #arbitrary number!
)

#names of the output object
names(distantia_kmeans)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"               

#kmeans object
distantia_kmeans$cluster_object
#> K-means clustering with 5 clusters of sizes 2, 2, 2, 1, 3
#> 
#> Cluster means:
#>      Alameda      Butte Contra_Costa  El_Dorado     Fresno  Humboldt  Imperial
#> 1 0.03542951 0.29562899   0.03542951 0.29016658 0.25722293 0.3490204 0.3731122
#> 2 0.30007870 0.25406472   0.26922153 0.30646700 0.06569086 0.4164782 0.2885995
#> 3 0.30828661 0.06977709   0.27750896 0.06977709 0.29914948 0.2163304 0.4896454
#> 4 0.35992954 0.23409632   0.33811131 0.19856448 0.42480715 0.0000000 0.5994240
#> 5 0.36102198 0.46538106   0.34824941 0.48497025 0.23063402 0.5741377 0.1618160
#>         Kern     Kings Los_Angeles
#> 1 0.31207730 0.4235817   0.2672132
#> 2 0.06569086 0.2731797   0.2616200
#> 3 0.26138223 0.5023917   0.4334898
#> 4 0.40814931 0.6088643   0.5141249
#> 5 0.31829880 0.1563507   0.1598843
#> 
#> Clustering vector:
#>      Alameda        Butte Contra_Costa    El_Dorado       Fresno     Humboldt 
#>            1            3            1            3            2            4 
#>     Imperial         Kern        Kings  Los_Angeles 
#>            5            2            5            5 
#> 
#> Within cluster sum of squares by cluster:
#> [1] 0.007575396 0.033688316 0.024263818 0.000000000 0.152614189
#>  (between_SS / total_SS =  89.4 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      

#distance matrix used for clustering
distantia_kmeans$d
#>                 Alameda     Butte Contra_Costa El_Dorado    Fresno  Humboldt
#> Alameda      0.00000000 0.3143751   0.07085902 0.3021981 0.2698072 0.3599295
#> Butte        0.31437509 0.0000000   0.27688289 0.1395542 0.2757251 0.2340963
#> Contra_Costa 0.07085902 0.2768829   0.00000000 0.2781350 0.2446387 0.3381113
#> El_Dorado    0.30219813 0.1395542   0.27813504 0.0000000 0.3225738 0.1985645
#> Fresno       0.26980716 0.2757251   0.24463870 0.3225738 0.0000000 0.4248072
#> Humboldt     0.35992954 0.2340963   0.33811131 0.1985645 0.4248072 0.0000000
#> Imperial     0.37624785 0.4801786   0.36997653 0.4991123 0.2395246 0.5994240
#> Kern         0.33035024 0.2324043   0.29380436 0.2903602 0.1313817 0.4081493
#> Kings        0.43556491 0.4789286   0.41159846 0.5258548 0.2252656 0.6088643
#> Los_Angeles  0.27125319 0.4370360   0.26317323 0.4299436 0.2271119 0.5141249
#>               Imperial      Kern     Kings Los_Angeles
#> Alameda      0.3762478 0.3303502 0.4355649   0.2712532
#> Butte        0.4801786 0.2324043 0.4789286   0.4370360
#> Contra_Costa 0.3699765 0.2938044 0.4115985   0.2631732
#> El_Dorado    0.4991123 0.2903602 0.5258548   0.4299436
#> Fresno       0.2395246 0.1313817 0.2252656   0.2271119
#> Humboldt     0.5994240 0.4081493 0.6088643   0.5141249
#> Imperial     0.0000000 0.3376745 0.2374235   0.2480245
#> Kern         0.3376745 0.0000000 0.3210939   0.2961280
#> Kings        0.2374235 0.3210939 0.0000000   0.2316285
#> Los_Angeles  0.2480245 0.2961280 0.2316285   0.0000000
#> attr(,"distantia_args")
#>    distance diagonal weighted ignore_blocks lock_step group
#> 1 euclidean     TRUE     TRUE         FALSE      TRUE     1
#> attr(,"type")
#> [1] "distantia_matrix"
#> attr(,"distance")
#> [1] "psi"

#number of clusters
distantia_kmeans$clusters
#> [1] 5

#clustering data frame
#group label in column "cluster"
distantia_kmeans$df
#>            name cluster silhouette_width
#> 1       Alameda       1        0.7638652
#> 2         Butte       3        0.4038600
#> 3  Contra_Costa       1        0.7368003
#> 4     El_Dorado       3        0.2971845
#> 5        Fresno       2        0.4303455
#> 6      Humboldt       4        0.0000000
#> 7      Imperial       5        0.1589591
#> 8          Kern       2        0.4973579
#> 9         Kings       5        0.1414957
#> 10  Los_Angeles       5        0.0833020

#mean silhouette width of the clustering solution
distantia_kmeans$silhouette_width
#> [1] 0.351317

#kmeans plot
# factoextra::fviz_cluster(
#   object = distantia_kmeans$cluster_object,
#   data = distantia_kmeans$d,
#   repel = TRUE
# )

#optimized kmeans
#---------------------------------
#auto-optimization of clusters and method
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = NULL
)

#names of the output object
#a new object named "optimization" should appear
names(distantia_kmeans)
#> [1] "cluster_object"   "clusters"         "silhouette_width" "df"              
#> [5] "d"                "optimization"    

#first rows of the optimization data frame
#optimized clustering in first row
head(distantia_kmeans$optimization)
#>   clusters silhouette_mean
#> 1        4       0.3953577
#> 2        5       0.3513170
#> 3        2       0.3435351
#> 4        6       0.3107771
#> 5        7       0.3084695
#> 6        3       0.2760140

#disable parallelization
future::plan(
  future::sequential
)