K-Means Clustering of Dissimilarity Analysis Data Frames

This function combines the dissimilarity scores computed by distantia(), the K-means clustering method implemented in stats::kmeans(), and the clustering optimization method implemented in utils_cluster_hclust_optimizer() to help group together time series with similar features.

When clusters = NULL, the function utils_cluster_hclust_optimizer() is run underneath to perform a parallelized grid search to find the number of clusters maximizing the overall silhouette width of the clustering solution (see utils_cluster_silhouette()). The grid search function supports parallelization via future::plan() and a progress bar generated by the progressr package (see Examples).

Usage

distantia_cluster_kmeans(df = NULL, clusters = NULL, seed = 1)

Arguments

df

(required, data frame) Output of distantia(). Default: NULL

clusters

(required, integer) Number of groups to generate. If NULL (default), utils_cluster_kmeans_optimizer() is used to find the number of clusters that maximizes the mean silhouette width of the clustering solution (see utils_cluster_silhouette()). Default: NULL

seed

(optional, integer) Random seed to be used during the K-means computation. Default: 1

Value

list:

cluster_object: kmeans object object for further analyses and custom plotting.

clusters: integer, number of clusters.

silhouette_width: mean silhouette width of the clustering solution.

df: data frame with time series names, their cluster label, and their individual silhouette width scores.

d: psi distance matrix used for clustering.

optimization: only if clusters = NULL, data frame with optimization results from utils_cluster_hclust_optimizer().

See also

Examples

#for large datasets, parallelization accelerates cluster optimization
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
# progressr::handlers(global = TRUE)

#daily covid prevalence in California counties
data("covid_prevalence")

#load as tsl and aggregate to monthly data to accelerate example execution
tsl <- tsl_initialize(
  x = covid_prevalence,
  id_column = "id",
  time_column = "time"
)
#> Error in utils_prepare_df(x = x, id_column = id_column, time_column = time_column): Column 'id' is not a column name of the data frame 'x'.

#subset 10 elements to accelerate example execution
tsl <- tsl_subset(
  tsl = tsl,
  names = 1:10
)
#> Error: object 'tsl' not found

#aggregateto monthly data to accelerate example execution
tsl <- tsl_aggregate(
  tsl = tsl,
  new_time = "months",
  fun = sum
)
#> Error: object 'tsl' not found

if(interactive()){
  #plotting first three time series
  tsl_plot(
    tsl = tsl[1:3],
    guide_columns = 3
    )
}

#dissimilarity analysis
distantia_df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)
#> Error: object 'tsl' not found

#kmeans with a given number of clusters
#-------------------------------------------------------
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = 5 #arbitrary number!
)
#> Error: object 'distantia_df' not found

#names of the output object
names(distantia_kmeans)
#> Error: object 'distantia_kmeans' not found

#kmeans object
distantia_kmeans$cluster_object
#> Error: object 'distantia_kmeans' not found

#distance matrix used for clustering
distantia_kmeans$d
#> Error: object 'distantia_kmeans' not found

#number of clusters
distantia_kmeans$clusters
#> Error: object 'distantia_kmeans' not found

#clustering data frame
#group label in column "cluster"
distantia_kmeans$df
#> Error: object 'distantia_kmeans' not found

#mean silhouette width of the clustering solution
distantia_kmeans$silhouette_width
#> Error: object 'distantia_kmeans' not found

#kmeans plot
# factoextra::fviz_cluster(
#   object = distantia_kmeans$cluster_object,
#   data = distantia_kmeans$d,
#   repel = TRUE
# )

#optimized kmeans
#---------------------------------
#auto-optimization of clusters and method
distantia_kmeans <- distantia_cluster_kmeans(
  df = distantia_df,
  clusters = NULL
)
#> Error: object 'distantia_df' not found

#names of the output object
#a new object named "optimization" should appear
names(distantia_kmeans)
#> Error: object 'distantia_kmeans' not found

#first rows of the optimization data frame
#optimized clustering in first row
head(distantia_kmeans$optimization)
#> Error: object 'distantia_kmeans' not found

#disable parallelization
future::plan(
  future::sequential
)