Dissimilarity Model Frame

This function generates a model frame for statistical or machine learning analysis from these objects:

: Dissimilarity data frame generated by distantia(). The output model frame will have as many rows as this data frame.

: Data frame with static descriptors of the time series. These descriptors are converted to distances between pairs of time series via distance_matrix().

: List defining new predictors as combinations of other existing predictors. This feature allows grouping together predictors that have a common meaning. For example, predictors_list = list(temperature = c("temperature_mean", "temperature_min", "temperature_max") generates a new predictor named "temperature", which results from computing the distances of the vector of temperature variables for each pair of time series. Predictors in one of such groups will be scaled before distance computation if their maximum standard deviations differ by a factor of 10 or more.

The resulting data frame contains the following columns:

x and y: names of the pair of time series represented in the row.

psi: dissimilarity between x and y.

predictors columns: representing the distance between the values of the given static predictor between x and y.

(optional) distance:If the static predictors data frame is an sf object, then this predictor is created via sf::st_distance().

Statistical or machine learning analyses based on this data frame may help uncover drivers of dissimilarity. Model coefficients or importance scores generated from this model frame represent the effect of the distance between predictors on the dissimilarity between time series.

This function supports a parallelization setup via future::plan(), and progress bars provided by the package progressr.

Usage

distantia_model_frame(
  df = NULL,
  predictors_df = NULL,
  predictors_list = NULL,
  predictors_scaled = FALSE,
  distance = "euclidean"
)

Arguments

df

(required, data frame) Output of distantia(), distantia_ls(), distantia_dtw(), or distantia_dtw_shift(). Default: NULL

predictors_df

(required, data frame) data frame with numeric predictors to be added to the model frame. Must have a column with the names in df$x and df$y. If sf data frame, the predictor "distance" is added to the model frame. Default: NULL

predictors_list

(optional, list) list defining new predictors as combinations of other predictors in predictors_df. For example, predictors_list = list(a = c("b", "c")) uses the columns "b" and "c" from predictors_df to generate the predictor a in the model frame. Default: NULL

predictors_scaled

(optional, logical) if TRUE, all predictors are scaled and centered with scale(). Default: FALSE

distance

(optional, character vector) name or abbreviation of the distance method. Valid values are in the columns "names" and "abbreviation" of the dataset distances. Default: "euclidean".

Value

data frame: with attributes "response", "predictors" and "formula".

See also

Examples

#covid prevalence in California counties
tsl <- tsl_initialize(
  x = covid_prevalence,
  name_column = "name",
  time_column = "time"
)

#dissimilarity analysis
df <- distantia(
  tsl = tsl,
  lock_step = TRUE
)

#combine several predictors
#into a new one
predictors_list <- list(
  economy = c(
    "poverty_percentage",
    "median_income",
    "domestic_product"
    )
)

#generate model frame
model_frame <- distantia_model_frame(
  df = df,
  predictors_df = covid_counties,
  predictors_list = predictors_list,
  predictors_scaled = TRUE
)

head(model_frame)
#>         psi   distance    economy area_hectares population poverty_percentage
#> 1 0.8726115 -1.1308600 -1.1122870    -0.7247975 -0.4884546         -1.1317715
#> 2 1.0656371 -1.1214496  0.4411267    -0.6664445 -0.1997738         -0.6664702
#> 3 1.1620553 -1.1308600 -1.1999329    -0.7391627 -0.4005782         -1.2480968
#> 4 1.2578125 -0.9213699 -0.6682476    -0.5909790 -0.1039143         -0.2884129
#> 5 1.2919255 -1.1308600 -1.0919021    -0.7182193 -0.5222018         -0.8991208
#> 6 1.2979351 -1.1308600  1.3447123    -0.5316916  0.1555335         -0.3465756
#>   median_income domestic_product daily_miles_traveled employed_percentage
#> 1    -0.9813874       -0.4910881            0.6672299           0.1688222
#> 2    -0.1807629        1.3397775           -0.9756356           0.6869774
#> 3    -1.2493713       -0.4668099           -0.6256303          -0.7943577
#> 4    -0.6378171       -0.6777708           -0.3619226          -0.2478307
#> 5    -0.9986465       -0.7233418           -0.9946619          -1.1991998
#> 6     1.3065864        1.9436225           -1.1501375          -0.4699586

#names of response and predictors
#and an additive formula
#are stored as attributes
attributes(model_frame)$response
#> [1] "psi"
attributes(model_frame)$predictors
#> [1] "distance"             "economy"              "area_hectares"       
#> [4] "population"           "poverty_percentage"   "median_income"       
#> [7] "domestic_product"     "daily_miles_traveled" "employed_percentage" 
attributes(model_frame)$formula
#> psi ~ distance + economy + area_hectares + population + poverty_percentage + 
#>     median_income + domestic_product + daily_miles_traveled + 
#>     employed_percentage
#> <environment: 0x55a8196f9720>

#linear model
model <- lm(
  formula = attributes(model_frame)$formula,
  data = model_frame
)

summary(model)
#> 
#> Call:
#> lm(formula = attributes(model_frame)$formula, data = model_frame)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -2.3579 -0.4988  0.1004  0.5309  1.6577 
#> 
#> Coefficients:
#>                       Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)           3.497370   0.029861 117.120  < 2e-16 ***
#> distance              0.213906   0.032047   6.675  5.5e-11 ***
#> economy               0.392631   0.260504   1.507   0.1323    
#> area_hectares         0.095988   0.030623   3.134   0.0018 ** 
#> population            0.057896   0.031454   1.841   0.0661 .  
#> poverty_percentage   -0.195852   0.119502  -1.639   0.1017    
#> median_income         0.031322   0.095096   0.329   0.7420    
#> domestic_product     -0.161710   0.155524  -1.040   0.2989    
#> daily_miles_traveled -0.009979   0.039144  -0.255   0.7989    
#> employed_percentage   0.115063   0.051094   2.252   0.0247 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.7495 on 620 degrees of freedom
#> Multiple R-squared:  0.1967,	Adjusted R-squared:  0.185 
#> F-statistic: 16.86 on 9 and 620 DF,  p-value: < 2.2e-16
#>