Split Dissimilarity Analysis Data Frames by Combinations of Arguments

Internal function to split a distantia data frame by groups of the arguments 'distance', 'diagonal', and 'lock_step'.

Usage

utils_distantia_df_split(df = NULL)

Arguments

df

(required, data frame) Output of distantia() or distantia_aggregate(). Default: NULL

Value

list

See also

Examples

#three time series
#climate and ndvi in Fagus sylvatica stands in Spain, Germany, and Sweden
tsl <- tsl_initialize(
  x = fagus_dynamics,
  name_column = "name",
  time_column = "time"
)

#dissimilarity analysis with four combinations of parameters
df <- distantia(
  tsl = tsl,
  distance = c(
    "euclidean",
    "manhattan"
    ),
  lock_step = c(
    TRUE,
    FALSE
  )
)

#split by combinations of parameters
df_split <- utils_distantia_df_split(
  df = df
)

#print output
df_split
#> $`1`
#>    distance diagonal lock_step       x      y bandwidth       psi group
#> 4 euclidean     TRUE     FALSE Germany Sweden         1 0.7285701     1
#> 5 euclidean     TRUE     FALSE Germany  Spain         1 1.0360268     1
#> 6 euclidean     TRUE     FALSE   Spain Sweden         1 1.0237508     1
#> 
#> $`2`
#>     distance diagonal lock_step       x      y bandwidth      psi group
#> 10 manhattan     TRUE     FALSE Germany  Spain         1 1.134012     2
#> 11 manhattan     TRUE     FALSE   Spain Sweden         1 1.166331     2
#> 12 manhattan     TRUE     FALSE Germany Sweden         1 0.761250     2
#> 
#> $`3`
#>    distance diagonal lock_step       x      y bandwidth       psi group
#> 7 manhattan       NA      TRUE Germany Sweden        NA 0.8271986     3
#> 8 manhattan       NA      TRUE   Spain Sweden        NA 1.5755002     3
#> 9 manhattan       NA      TRUE Germany  Spain        NA 1.3737381     3
#> 
#> $`4`
#>    distance diagonal lock_step       x      y bandwidth       psi group
#> 1 euclidean       NA      TRUE Germany  Spain        NA 1.3962157     4
#> 2 euclidean       NA      TRUE   Spain Sweden        NA 1.5703931     4
#> 3 euclidean       NA      TRUE Germany Sweden        NA 0.8364652     4
#> 

#class and length of the output
class(df_split)
#> [1] "list"
length(df_split)
#> [1] 4