Data Frame with Contribution of Individual Variables to Dissimilarity to Wide Format

Takes as input a data frame returned by distantia_importance() to return a data frame with one pair of time series per row, and the following columns:

most_similar: name of the variable with the highest contribution to similarity (most negative value in the importance column) for each pair of time series.

most_dissimilar: name of the variable with the highest contribution to dissimilarity (most positive value in the importance column) for each pair of time series.

importance__variable_name: contribution to similarity (negative values) or dissimilarity (positive values) of the given variable.

psi_only_with__variable_name: dissimilarity of the two time series when only using the given variable.

psi_without__variable_name: dissimilarity of the two time series when removing the given variable.

Usage

utils_importance_df_to_wide(df = NULL, sep = "__")

Arguments

df

(required, data frame) Output of distantia_importance(). Default: NULL

sep

(required, character string) Separator between the name of the importance metric and the time series variable. Default: "__".

Value

data frame

See also

Examples

#prepare time series
data("fagus_dynamics")

tsl <- tsl_initialize(
  x = fagus_dynamics,
  name_column = "name",
  time_column = "time"
) |>
  tsl_transform(
    f = f_scale
  )
#> distantia::tsl_diagnose(): No issues found.

#importance data frame
df <- distantia_importance(
  tsl = tsl
)
#> distantia::tsl_diagnose(): No issues found.
df
#>         x      y       psi    variable importance               effect
#> 1 Germany  Spain 0.9947403         evi -14.508771 increases similarity
#> 2 Germany  Spain 0.9947403    rainfall  35.666900 decreases similarity
#> 3 Germany  Spain 0.9947403 temperature -42.473911 increases similarity
#> 4 Germany Sweden 0.7018922         evi  -6.523915 increases similarity
#> 5 Germany Sweden 0.7018922    rainfall  36.531485 decreases similarity
#> 6 Germany Sweden 0.7018922 temperature -54.420090 increases similarity
#> 7   Spain Sweden 1.1082660         evi -36.569755 increases similarity
#> 8   Spain Sweden 1.1082660    rainfall  57.678064 decreases similarity
#> 9   Spain Sweden 1.1082660 temperature -50.150369 increases similarity
#>   psi_difference psi_without psi_only_with  distance diagonal weighted
#> 1    -0.14432459   1.0193192     0.8749946 euclidean     TRUE     TRUE
#> 2     0.35479302   0.7900431     1.1448361 euclidean     TRUE     TRUE
#> 3    -0.42250509   1.0646350     0.6421299 euclidean     TRUE     TRUE
#> 4    -0.04579085   0.7065423     0.6607514 euclidean     TRUE     TRUE
#> 5     0.25641166   0.5474000     0.8038117 euclidean     TRUE     TRUE
#> 6    -0.38197039   0.7608013     0.3788309 euclidean     TRUE     TRUE
#> 7    -0.40529016   1.1806979     0.7754078 euclidean     TRUE     TRUE
#> 8     0.63922638   0.7313681     1.3705945 euclidean     TRUE     TRUE
#> 9    -0.55579950   1.2030043     0.6472048 euclidean     TRUE     TRUE
#>   ignore_blocks lock_step robust
#> 1         FALSE     FALSE   TRUE
#> 2         FALSE     FALSE   TRUE
#> 3         FALSE     FALSE   TRUE
#> 4         FALSE     FALSE   TRUE
#> 5         FALSE     FALSE   TRUE
#> 6         FALSE     FALSE   TRUE
#> 7         FALSE     FALSE   TRUE
#> 8         FALSE     FALSE   TRUE
#> 9         FALSE     FALSE   TRUE

#to wide format
df_wide <- utils_importance_df_to_wide(
  df = df
)
df_wide
#>         x      y       psi most_similar most_dissimilar importance__evi
#> 1 Germany  Spain 0.9947403  temperature        rainfall      -14.508771
#> 2 Germany Sweden 0.7018922  temperature        rainfall       -6.523915
#> 3   Spain Sweden 1.1082660  temperature        rainfall      -36.569755
#>   importance__rainfall importance__temperature psi_only_with__evi
#> 1             35.66690               -42.47391          0.8749946
#> 2             36.53148               -54.42009          0.6607514
#> 3             57.67806               -50.15037          0.7754078
#>   psi_only_with__rainfall psi_only_with__temperature psi_without__evi
#> 1               1.1448361                  0.6421299        1.0193192
#> 2               0.8038117                  0.3788309        0.7065423
#> 3               1.3705945                  0.6472048        1.1806979
#>   psi_without__rainfall psi_without__temperature
#> 1             0.7900431                1.0646350
#> 2             0.5474000                0.7608013
#> 3             0.7313681                1.2030043