Computes the dissimilarity measure psi on two or more sequences.

This workflow executes the following steps:

Computes the autosum of the sequences with autoSum.

Computes the distance matrix with distanceMatrix.

Uses the distance matrix to compute the least cost matrix with leastCostMatrix.

Extracts the cost of the least cost path with leastCost.

Computes the dissimilarity measure psi with the function psi.

Delivers an output of type "list" (default), "data.frame" or "matrix", depending on the user input, through formatPsi.

workflowPsi(
  sequences = NULL,
  grouping.column = NULL,
  time.column = NULL,
  exclude.columns = NULL,
  method = "manhattan",
  diagonal = FALSE,
  format = "dataframe",
  parallel.execution = TRUE
  )

Arguments

sequences
dataframe with multiple sequences identified by a grouping column generated by prepareSequences.

grouping.column
character string, name of the column in sequences to be used to identify separates sequences within the file.

time.column
character string, name of the column with time/depth/rank data.

exclude.columns
character string or character vector with column names in sequences to be excluded from the analysis.

method
character string naming a distance metric. Valid entries are: "manhattan", "euclidean", "chi", and "hellinger". Invalid entries will throw an error.

diagonal
boolean, if TRUE, diagonals are included in the computation of the least cost path. Defaults to FALSE, as the original algorithm did not include diagonals in the computation of the least cost path.

format
string, type of output. One of: "data.frame", "matrix". If NULL or empty, a list is returned.

parallel.execution
boolean, if TRUE (default), execution is parallelized, and serialized if FALSE.

Value

A list, matrix, or dataframe, with sequence names and psi values.

Examples

# NOT RUN {
data("sequencesMIS")
#prepare sequences
MIS.sequences <- prepareSequences(
  sequences = sequencesMIS,
  grouping.column = "MIS",
  if.empty.cases = "zero",
  transformation = "hellinger"
  )

#execute workflow to compute psi
MIS.psi <- workflowPsi(
 sequences = MIS.sequences,
 grouping.column = "MIS",
 time.column = NULL,
 exclude.columns = NULL,
 method = "manhattan",
 diagonal = FALSE,
 output = "dataframe"
 )

MIS.psi

# }