A refactored version of workflowPsi with a higher performance (hence the suffix HP).

Ideal for large analyses with hundreds to thousands of sequences. Several options available in workflowPsi have been removed from this function in order to simplify the code as much as possible. Psi is computed with the options diagonal = TRUE, ignore.blocks = TRUE, and method = "euclidean".

workflowPsiHP(
  sequences = NULL,
  grouping.column = NULL,
  time.column = NULL,
  exclude.columns = NULL,
  parallel.execution = TRUE
  )

Arguments

sequences
dataframe with multiple sequences identified by a grouping column generated by prepareSequences.

grouping.column
character string, name of the column in sequences to be used to identify separates sequences within the file.

time.column
character string, name of the column with time/depth/rank data.

exclude.columns
character string or character vector with column names in sequences to be excluded from the analysis.

parallel.execution
boolean, if TRUE (default), execution is parallelized, and serialized if FALSE.

Value

A dataframe with sequence names and psi values.

Details

Due to limitations of the function permutations, the maximum number of groups (according to grouping.column) is around 30000. Besides, a combinations table of this size takes, roughlyl, 7GB of memory.

Examples

data("sequencesMIS")
#prepare sequences
MIS.sequences <- prepareSequences(
  sequences = sequencesMIS,
  grouping.column = "MIS",
  if.empty.cases = "zero",
  transformation = "hellinger"
  )

#execute workflow to compute psi
MIS.psi <- workflowPsiHP(
 sequences = MIS.sequences,
 grouping.column = "MIS",
 parallel.execution = FALSE
 )

MIS.psi

#>         A      B       psi
#> 1   MIS-1  MIS-2 2.2943523
#> 2   MIS-1  MIS-3 2.3166569
#> 3   MIS-1  MIS-4 1.8731964
#> 4   MIS-1  MIS-5 1.7265078
#> 5   MIS-1  MIS-6 2.3369391
#> 6   MIS-1  MIS-7 1.7711805
#> 7   MIS-1  MIS-8 2.2918180
#> 8   MIS-1  MIS-9 2.2154998
#> 9   MIS-1 MIS-10 2.6994681
#> 10  MIS-1 MIS-11 2.3638545
#> 11  MIS-1 MIS-12 2.0586868
#> 12  MIS-2  MIS-3 1.1365382
#> 13  MIS-2  MIS-4 1.1214264
#> 14  MIS-2  MIS-5 1.6315633
#> 15  MIS-2  MIS-6 1.1273430
#> 16  MIS-2  MIS-7 2.1501248
#> 17  MIS-2  MIS-8 1.5455081
#> 18  MIS-2  MIS-9 1.6932739
#> 19  MIS-2 MIS-10 1.2975746
#> 20  MIS-2 MIS-11 1.9078618
#> 21  MIS-2 MIS-12 1.1002146
#> 22  MIS-3  MIS-4 1.2541221
#> 23  MIS-3  MIS-5 1.2467539
#> 24  MIS-3  MIS-6 0.8631049
#> 25  MIS-3  MIS-7 1.4692465
#> 26  MIS-3  MIS-8 1.0845708
#> 27  MIS-3  MIS-9 1.1724508
#> 28  MIS-3 MIS-10 1.0340321
#> 29  MIS-3 MIS-11 1.2770386
#> 30  MIS-3 MIS-12 0.9358788
#> 31  MIS-4  MIS-5 2.4810112
#> 32  MIS-4  MIS-6 1.1020176
#> 33  MIS-4  MIS-7 1.9632741
#> 34  MIS-4  MIS-8 1.5710346
#> 35  MIS-4  MIS-9 2.4481557
#> 36  MIS-4 MIS-10 1.6486674
#> 37  MIS-4 MIS-11 2.3919441
#> 38  MIS-4 MIS-12 1.5184440
#> 39  MIS-5  MIS-6 1.4883563
#> 40  MIS-5  MIS-7 0.9865546
#> 41  MIS-5  MIS-8 1.2889531
#> 42  MIS-5  MIS-9 1.0391586
#> 43  MIS-5 MIS-10 1.0174326
#> 44  MIS-5 MIS-11 1.2655812
#> 45  MIS-5 MIS-12 1.0241509
#> 46  MIS-6  MIS-7 1.6410644
#> 47  MIS-6  MIS-8 1.1928909
#> 48  MIS-6  MIS-9 1.3699400
#> 49  MIS-6 MIS-10 1.1669777
#> 50  MIS-6 MIS-11 1.4274162
#> 51  MIS-6 MIS-12 1.0735763
#> 52  MIS-7  MIS-8 1.0970102
#> 53  MIS-7  MIS-9 1.0597719
#> 54  MIS-7 MIS-10 1.1209314
#> 55  MIS-7 MIS-11 1.3568246
#> 56  MIS-7 MIS-12 1.0422664
#> 57  MIS-8  MIS-9 1.1982089
#> 58  MIS-8 MIS-10 1.1185626
#> 59  MIS-8 MIS-11 0.8642958
#> 60  MIS-8 MIS-12 1.0078508
#> 61  MIS-9 MIS-10 0.9421708
#> 62  MIS-9 MIS-11 1.0008698
#> 63  MIS-9 MIS-12 1.0530892
#> 64 MIS-10 MIS-11 0.9930614
#> 65 MIS-10 MIS-12 0.9099192
#> 66 MIS-11 MIS-12 0.9815528